Restructuration of the main file and fixing bug on rspec
