+++ .rspec
@@ -1,1 +1,3 @@
 --require spec_helper
+--color
+--format documentation
