+++ lib/player.rb
@@ -1,0 +1,21 @@
+
+class Player
+  attr_accessor :name, :symbol
+
+  def name=(name)
+    @name = name
+  end
+
+  def symbol=(symbol)
+    @symbol = symbol
+  end
+
+  def symbol
+    @symbol
+  end
+
+  def name
+    @name
+  end
+
+end
